requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
burnysc2==5.0.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
idna==2.10
importlib-metadata==2.1.0
loguru==0.5.3
mpyq==0.2.5
multidict==5.0.2
numpy==1.19.3
Pillow==7.2.0
pipenv==2020.8.13
portpicker==1.3.1
protobuf==3.14.0
pyglet==1.5.11
s2clientprotocol==5.0.4.82457.0
scipy==1.5.4
six==1.15.0
snakeviz==2.1.0
tornado==6.0.4
typing-extensions==3.7.4.3
virtualenv==20.0.33
virtualenv-clone==0.5.4
win32-setctime==1.0.3
yarl==1.6.3
zipp==3.4.0
